Add new events dispatching
